# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Question Answering API
  description: An API for adding knowledge and asking questions using a RAG (Retrieval Augmented Generation) system
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /add:
    post:
      summary: Add knowledge to the system
      description: Adds new information to the knowledge base for future question answering
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: The text content to add to the knowledge base
              required:
                - content
      responses:
        '200':
          description: Knowledge base updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Updated the knowledge base successfully."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: Error message

  /ask:
    post:
      summary: Ask a question
      description: Ask a question and get an answer based on the knowledge base and chat history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    description: The AI-generated answer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: Error message

components:
  schemas:
    Message:
      type: object
      properties:
        role:
          type: string
          description: The role of the message sender (e.g., "user" or "assistant")
        content:
          type: string
          description: The content of the message
          default: ""
      required:
        - role

    ConversationRequest:
      type: object
      properties:
        message:
          type: string
          description: The current question or message from the user
        chat_history:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: The history of previous messages in the conversation
      required:
        - message
        - chat_history